+++ src/CrudPanel.php
@@ -22,11 +22,12 @@ use Backpack\CRUD\PanelTraits\AutoFocus;
 use Backpack\CRUD\PanelTraits\FakeFields;
 use Backpack\CRUD\PanelTraits\FakeColumns;
 use Illuminate\Database\Eloquent\Collection;
+use Backpack\CRUD\PanelTraits\RequiredFields;
 use Backpack\CRUD\PanelTraits\ViewsAndRestoresRevisions;
 
 class CrudPanel
 {
-    use Create, Read, Search, Update, Delete, Errors, Reorder, Access, Columns, Fields, Query, Buttons, AutoSet, FakeFields, FakeColumns, ViewsAndRestoresRevisions, AutoFocus, Filters, Tabs, Views;
+    use Create, Read, Search, Update, Delete, Errors, Reorder, Access, Columns, Fields, Query, Buttons, AutoSet, FakeFields, FakeColumns, ViewsAndRestoresRevisions, AutoFocus, Filters, Tabs, Views, RequiredFields;
 
     // --------------
     // CRUD variables
@@ -172,6 +173,56 @@ class CrudPanel
     {
         $this->entity_name = $singular;
         $this->entity_name_plural = $plural;
+    }
+
+    // -----------------------------------------------
+    // ACTIONS - the current operation being processed
+    // -----------------------------------------------
+
+    /**
+     * Get the action being performed by the controller,
+     * including middleware names, route name, method name,
+     * namespace, prefix, etc.
+     *
+     * @return string The EntityCrudController route action array.
+     */
+    public function getAction()
+    {
+        return $this->request->route()->getAction();
+    }
+
+    /**
+     * Get the full name of the controller method
+     * currently being called (including namespace).
+     *
+     * @return string The EntityCrudController full method name with namespace.
+     */
+    public function getActionName()
+    {
+        return $this->request->route()->getActionName();
+    }
+
+    /**
+     * Get the name of the controller method
+     * currently being called.
+     *
+     * @return string The EntityCrudController method name.
+     */
+    public function getActionMethod()
+    {
+        return $this->request->route()->getActionMethod();
+    }
+
+    /**
+     * Check if the controller method being called
+     * matches a given string.
+     *
+     * @param  string $methodName   Name of the method (ex: index, create, update)
+     * @return bool                 Whether the condition is met or not.
+     */
+    public function actionIs($methodName)
+    {
+        return $methodName === $this->getActionMethod();
     }
 
     // ----------------------------------